rename config_instruction to instruction
